Belethor's General Mods v0.0.1-2 Common.Schema.Mod

Store information about a mod

Attribute	Default Value	Meaning

name	no default	the mods name

desc	"todo: add description"	the mods description

published	false	wether the mod should be publicly accessibly

image	no_image.jpg	an url to an image

sse	nil	a number of links for its Skyrim Special Edition implementation

oldrim	nil	a number of links for the OG Skyrim Engine released in 11.11.2011

tags	[]	the number of tags

Types

change()

change() :: %{
  optional(:name) => binary(),
  optional(:desc) => binary(),
  optional(:published) => binary(),
  optional(:image) => binary(),
  optional(:sse) => Common.Schema.ModFile.change(),
  optional(:oldrim) => Common.Schema.ModFile.change(),
  optional(:tags) => tag_change()
}

id()

id() :: term()

t()

t() :: %Common.Schema.Mod{
  __meta__: term(),
  desc: String.t(),
  id: id(),
  image: String.t(),
  name: String.t(),
  oldrim: Common.Schema.ModFile.t() | nil,
  published: bool(),
  sse: Common.Schema.ModFile.t() | nil,
  tags: [Common.Schema.ModTag.t()] | Ecto.Association.NotLoaded.t()
}

tag_change()

tag_change() :: [
  add: Common.Schema.ModTag.t(),
  delete: Common.Schema.ModTag.t()
]

Functions

clean_changes(unclean)

clean_changes(unclean :: Common.unclean_change()) :: change()

if you want to to change the tags, this won't be enough.

clean_tag_changes(unclean)

clean_tag_changes(unclean :: Common.unclean_change()) :: tag_change()

new(changes \\ %{})

new(changes :: change()) :: Ecto.Changeset.t()

Creates a new %Mod to be fed into Repo.insert!/2 for example. Sets default values, if they are not present in the change argument.

preload(mod)

preload(mod :: t()) :: [Common.Schema.ModTag.t()]

Loads data, from other data. mods.tags is filled, with that.