Belethor's General Mods v0.0.1-2 Common.Schema.ModFile

Store information about a game-version-specific file from a mod

Attribute	Meaning

nexus	a link to the files on nexusmods.com

steam	a link to the files on the steam workshop

bethedsa	a link to the bethesda mod store

console_compat	a boolean, if this mod can run on a console

Types

t()

t() :: %Common.Schema.ModFile{
  bethesda: String.t(),
  console_compat: bool(),
  id: term(),
  nexus: String.t(),
  steam: String.t()
}

Functions

changeset(modfile, changes)