Belethor's General Mods v0.0.1-1 Crawler.Client behaviour

This module describes client behaviour to be used by Crawler.TaskManager.search/4.

The only function to be implemented is the search/1 function.

Types

args()

args() :: [any()]

TODO this should be done with the first real impls of clients

result()

result() :: {:ok, [Common.Struct.Mod.t()]} | {:error, term()}

Callbacks

search(args)

search(args()) :: result()

This function is meant to query an provider. In the usual version of an http based api from the provider, you need to do: Query the provider server, with the argumens of in the query(). Then parse the data into [result()]. If all this worked as expected return {:ok, [result()]}. If not return {:error, reason :: term()}.