Belethor's General Mods v0.0.1-1 Crawler.Provider behaviour

behaviour to implemented a provider. use this module to implement the callbacks the standard way.

If you use this module, a Crawler.TaskManager is expected run as __YOUR_MODULE__.TaskManager and a Crawler.Client should be implemented at __YOUR_MODULE__.Client

Callbacks

search(arg0)

search(Crawler.Client.args()) :: Crawler.Client.result()

do a remote search, may return a error tuple

search!(arg0)

search!(Crawler.Client.args()) :: [Common.Struct.Mod.t()]

do a remote search, throws on error