Belethor's General Mods v0.0.1-1 Crawler.TaskManager

Module to limit the amount of concurrent running task. If the task maximum is reached, the request will be queued.

Types

max()

max() :: pos_integer() | :infinty

start_options()

start_options() ::
  {:max, max()}
  | {:task_supervisor, Supervisor.supervisor()}
  | GenServer.options()

start options for starting up a TaskManager

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

search(query, manager, client, timeout \\ 5000)

search(
  query :: Crawler.Client.args(),
  manager :: GenServer.name(),
  client :: module(),
  timeout()
) :: Crawler.Client.result()

execute client.search(query) in a rate limited way.

the search callback is defined in Crawler.Client.

start_link(opts)

start_link(opts :: [start_options()]) :: GenServer.on_start()

start an instance

:max and :task_supervisor are required:

:max the maximum tasks running in parrallel

:task_manager the Task.Supervisor to supervise the tasks.

GenServer.options() can be added, but are optional.